Backend/requirements.txt
asgiref==3.8.1
blinker==1.8.2
certifi==2025.4.26
charset-normalizer==3.4.2
click==8.1.7
colorama==0.4.6
crispy-bootstrap4==2024.1
crispy-bootstrap5==2024.2
distlib==0.3.8
Django==5.0.6
django-bootstrap5==24.2
django-cors-headers==4.7.0
django-crispy-forms==2.2
django-filter==24.3
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.0
drf-nested-routers==0.94.2
filelock==3.15.4
Flask==3.0.3
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.4
joblib==1.4.2
lxml==5.3.0
MarkupSafe==3.0.2
numpy==2.0.0
pandas==2.2.2
pillow==10.4.0
platformdirs==4.2.2
PyJWT==2.9.0
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
python-pptx==1.0.2
pytz==2024.1
requests==2.32.3
six==1.16.0
sqlparse==0.5.0
typing_extensions==4.12.2
tzdata==2024.1
urllib3==2.4.0
virtualenv==20.26.3
Werkzeug==3.0.4
XlsxWriter==3.2.1
